+++ server/db/update_scripts/31_update_v6.16.0_v7.0.0.sql
@@ -578,7 +578,7 @@ DELIMITER ;
 
 -- add field `formName` to style `showUserInput`; Set the data from `source` field to the formName and remove the `source` field
 -- insert field `formName` in style `showUserInput`
-INSERT IGNORE INTO `styles_fields` (`id_styles`, `id_fields`, `default_value`, `help`)  VALUES (get_style_id('showUserInput'), get_field_id('formName'), '', 'Select a dataTable that will be loaded and show the user data entries.');
+INSERT IGNORE INTO `styles_fields` (`id_styles`, `id_fields`, `default_value`, `help`)  VALUES (get_style_id('showUserInput'), get_field_id('formName'), '', 'Select a data table that will be loaded and show the user data entries.');
 -- link the source field to the dataTables
 UPDATE sections_fields_translation sft
 INNER JOIN sections s ON (s.id = sft.id_sections)
@@ -611,6 +611,10 @@ WHERE `name` = 'formName';
 UPDATE fieldType
 SET `name` = 'select-data_table'
 WHERE `name` = 'select-formName';
+
+UPDATE styles_fields
+SET `help` = 'Select a data tabe which will be linked to the style'
+WHERE id_fields = get_field_id('data_table') AND id_styles IN (get_style_id('entryRecord'), get_style_id('entryList'));